+++ bytebank/src/main/kotlin/br/com/alura/bytebank/modelo/Endereco.kt
@@ -48,5 +48,11 @@ class Endereco(
         return result
     }
 
-
+    fun completo(): String {
+        return """
+$logradouro - $numero, $bairro, $cidade - $estado
+$cep
+$complemento            
+""".trimIndent()
+    }
 }
