+++ bytebank/src/main/kotlin/br/com/alura/bytebank/teste/testaHOF.kt
@@ -1,0 +1,23 @@
+package br.com.alura.bytebank.teste
+
+import br.com.alura.bytebank.modelo.Autenticavel
+import br.com.alura.bytebank.modelo.SistemaInterno
+
+fun testaHOF() {
+    soma(1, 5, resultado = (::println))
+
+    val autenticavel = object : Autenticavel {
+        val senha = 1234
+        override fun autentica(senha: Int) = this.senha == senha
+    }
+
+    SistemaInterno().entra(autenticavel, 1234, autenticado = {
+        println("realizar pagamento")
+    })
+}
+
+fun soma(a: Int, b: Int, resultado: (Int) -> Unit) {
+    println("antes da soma")
+    resultado(a + b)
+    println("depois da soma")
+}
